+++ spec/graphql/support/test_schema.rb
@@ -2,8 +2,10 @@
 
 # rubocop:disable Metrics/MethodLength
 def build_test_schema(options = {})
+  test_tracer = options.delete(:tracer)
   schema = Class.new(GraphQL::Schema) do
     use GraphQL::PersistedQueries, options
+    tracer test_tracer if test_tracer
 
     query(
       Class.new(GraphQL::Schema::Object) do